fix(dress): address review — break circular import, add batch logging
- Move describe_priority_context from dress.py to dress_context.py to
  eliminate the runtime circular import (dress_context → dress)
- Add log.warning when a brief batch returns None (was silently skipped)
- Add comment explaining composition log is best-effort across batches
- Fix test mock to parse passage IDs from context instead of using a
  fragile call_count-based index
